refactor: custom hooks
