+++ src/Components/ProductCard/ProductCard.jsx
@@ -3,14 +3,13 @@ import { AddToCart, ToggleWishList } from "../";
 
 export const ProductCard = ({ product }) => {
   const navigate = useNavigate();
-
   const inStock = product.stock > 0;
 
   return (
-    <div className="box-shadow flex flex-col cur-point pos-rel product-card p-2">
+    <div className="box-shadow flex-col cur-point pos-rel product-card p-2">
       <div
         onClick={() => navigate(`/product/${product._id}`)}
-        className="w-100 p-4 flex flex-col"
+        className="w-100 p-4 flex-col"
       >
         <div className="flex-grow pos-rel">
           <img
@@ -21,7 +20,7 @@ export const ProductCard = ({ product }) => {
             loading="lazy"
           />
           {!inStock && (
-            <div className="pos-abs top w-100 h-100 flex justify-center align-center">
+            <div className="pos-abs top w-100 h-100 flex place-center">
               <div className="w-100 text-center p-2 bg-color-2 text-bold color-1">
                 Out of stock!
               </div>
@@ -32,7 +31,7 @@ export const ProductCard = ({ product }) => {
       </div>
       <div className="text-bold m-h-4 flex-grow">₹{product.price}</div>
 
-      <div className="flex justify-center align-center w-100">
+      <div className="flex place-center w-100">
         <AddToCart _id={product._id} inStock={inStock} />
       </div>
