+++ src/Components/ProductCard/ProductCardHorizontal.jsx
@@ -1,111 +1,35 @@
-import { useAuth, useUserData } from "../../hooks";
-import axios from "axios";
-import { BASE_URL } from "../../utils/utils";
+import { useUserActions } from "../../hooks";
+import { useState } from "react";
 
 export const ProductCardHorizontal = ({ product }) => {
-  const { userProfile, isUserLoggedIn } = useAuth();
-  const { userData, userDispatch } = useUserData();
+  const [loading, setLoading] = useState(false);
+  const {
+    decrementQuantity,
+    incrementQuantity,
+    removeFromCartOnClick,
+    isAlreadyInWishList,
+    addToWishList,
+  } = useUserActions(setLoading);
 
-  const decrementQuantity = async (_id, quantity) => {
-    if (isUserLoggedIn && userProfile?._id) {
-      const res = await axios.post(
-        `${BASE_URL}/cart/${userProfile._id}/${_id}`,
-        {
-          quantity: quantity - 1,
-        }
-      );
-
-      if (res.data.success) {
-        userDispatch({
-          type: "DECREMENT_CART",
-          payload: {
-            product: _id,
-          },
-        });
-      }
-      return;
+  const moveToWishListOnClick = () => {
+    if (!isAlreadyInWishList(product.product._id)) {
+      addToWishList(product.product._id);
     }
-  };
-
-  const incrementQuantity = async (_id, quantity) => {
-    if (isUserLoggedIn && userProfile?._id) {
-      const res = await axios.post(
-        `${BASE_URL}/cart/${userProfile._id}/${_id}`,
-        {
-          quantity: quantity + 1,
-        }
-      );
-
-      if (res.data.success) {
-        userDispatch({
-          type: "INCREMENT_CART",
-          payload: {
-            product: _id,
-          },
-        });
-      }
-      return;
-    }
-  };
-
-  const removeFromCartOnClick = async (_id) => {
-    if (isUserLoggedIn && userProfile?._id) {
-      const res = await axios.delete(
-        `${BASE_URL}/cart/${userProfile._id}/${_id}`
-      );
-
-      if (res.data.success) {
-        userDispatch({
-          type: "REMOVE_FROM_CART",
-          payload: {
-            product: _id,
-          },
-        });
-      }
-      return;
-    }
-  };
-
-  const isAlreadyInWishList = (_id) =>
-    userData.wishList.some((product) => product === _id);
-
-  const addToWishList = async (_id) => {
-    if (isUserLoggedIn && userProfile?._id && !isAlreadyInWishList(_id)) {
-      const res = await axios.post(`${BASE_URL}/wish-list/${userProfile._id}`, {
-        product: {
-          _id,
-        },
-      });
-
-      if (res.data.success) {
-        userDispatch({
-          type: "ADD_TO_WISHLIST",
-          payload: {
-            product: _id,
-          },
-        });
-      }
-
-      return;
-    }
+    removeFromCartOnClick(product.product._id);
   };
 
   return (
-    <div className="box-shadow flex align-center pos-rel product-card-h">
-      <div className="h-4 w-3 p-4 flex flex-col pos-rel">
+    <div className="box-shadow flex align-center pos-rel product-card-h m-v-2">
+      <div className="h-4 w-5 p-2 flex-col align-center space-around">
         <img
-          className="img-responsive"
-          style={{ width: "auto" }}
+          className="img-responsive w-auto"
           src={product.product.imageUrl}
           alt={product.product.name}
-          loading="lazy"
         />
         <button
-          onClick={() => {
-            addToWishList(product.product._id);
-            removeFromCartOnClick(product.product._id);
-          }}
-          className="btn btn-small bg-color-2 font-3 pos-abs bottom m-2 p-h-2 m-h-3 border-1"
+          disabled={loading}
+          onClick={moveToWishListOnClick}
+          className="btn bg-color-2 font-3 p-v-2 p-h-3"
         >
           Move to wishlist
         </button>
@@ -120,8 +44,8 @@ export const ProductCardHorizontal = ({ product }) => {
         <div className="flex space-even m-v-2">
           <div>Quantity: </div>
           <button
-            disabled={product.quantity === 1}
-            className="btn bg-inherit btn-decrease"
+            disabled={product.quantity === 1 || loading}
+            className="btn bg-inherit icon"
             onClick={() =>
               decrementQuantity(product.product._id, product.quantity)
             }
@@ -130,10 +54,11 @@ export const ProductCardHorizontal = ({ product }) => {
           </button>
           <div className="font-4">{product.quantity}</div>
           <button
+            disabled={loading}
             onClick={() =>
               incrementQuantity(product.product._id, product.quantity)
             }
-            className="btn bg-inherit"
+            className="btn bg-inherit icon"
           >
             <span className="material-icons-round">add</span>
           </button>
@@ -141,7 +66,8 @@ export const ProductCardHorizontal = ({ product }) => {
 
         <div className="pos-abs top-right">
           <button
-            className="btn bg-inherit pos-abs top-right m-2"
+            disabled={loading}
+            className="btn bg-inherit pos-abs top-right m-2 icon"
             onClick={() => removeFromCartOnClick(product.product._id)}
           >
             <span className="material-icons-round font-7 color-1">
