+++ src/hooks/index.js
@@ -5,3 +5,4 @@ export { useAxios } from "./useAxios";
 export { useScrollToTop } from "./useScrollToTop";
 export { useAuthPersist } from "./useAuthPersist";
 export { useQueryParams } from "./useQueryParams";
+export { useUserActions } from "./useUserActions";
