+++ src/pages/Addresses/Addresses.jsx
@@ -2,7 +2,7 @@
 export const Addresses = () => {
     return (
         <div>
-            This is addresses
+            This is addresses page.
         </div>
     )
 }
