+++ src/pages/ProductInfo/ProductInfo.jsx
@@ -2,18 +2,21 @@ import axios from "axios";
 import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { BASE_URL } from "../../utils/utils";
-import { useUserData, useAuth } from "../../hooks";
+import { useUserActions } from "../../hooks";
 import { Loading } from "../../Components";
 
 export const ProductInfo = () => {
-  const { id } = useParams();
   const [product, setProduct] = useState({});
-  const navigate = useNavigate();
   const [loading, setLoading] = useState(false);
-  const { userData, userDispatch } = useUserData();
-  const { userProfile, isUserLoggedIn } = useAuth();
-
-  const inStock = product?.stock > 0;
+  const navigate = useNavigate();
+  const { id } = useParams();
+  const {
+    addToCartOnClick,
+    addToWishList,
+    isAlreadyInCart,
+    isAlreadyInWishList,
+    removeFromWishList,
+  } = useUserActions(setLoading);
 
   useEffect(() => {
     window.scrollTo(0, 0);
@@ -25,92 +28,10 @@ export const ProductInfo = () => {
     })();
   }, []);
 
-  const addToWishList = async (_id) => {
-    if (isUserLoggedIn && userProfile?._id) {
-      setLoading(true);
-      const res = await axios.post(`${BASE_URL}/wish-list/${userProfile._id}`, {
-        product: {
-          _id,
-        },
-      });
-
-      if (res.data.success) {
-        userDispatch({
-          type: "ADD_TO_WISHLIST",
-          payload: {
-            product: _id,
-          },
-        });
-      }
-      setLoading(false);
-      return;
-    }
-    navigate("/login", {
-      state: {
-        message: "Login to add to wishlist.",
-        addTo: "WISHLIST",
-        productId: _id,
-      },
-    });
-  };
-
-  const isAlreadyInWishList = (_id) =>
-    userData.wishList.some((product) => product === _id);
-
-  const removeFromWishList = async (_id) => {
-    if (isUserLoggedIn && userProfile?._id) {
-      setLoading(true);
-      const res = await axios.delete(
-        `${BASE_URL}/wish-list/${userProfile._id}/${_id}`
-      );
-
-      if (res.data.success) {
-        userDispatch({
-          type: "REMOVE_FROM_WISHLIST",
-          payload: {
-            product: _id,
-          },
-        });
-      }
-      setLoading(false);
-    }
-  };
-
-  const addToCartOnClick = async (_id) => {
-    if (isUserLoggedIn && userProfile?._id) {
-      setLoading(true);
-      const res = await axios.post(`${BASE_URL}/cart/${userProfile._id}`, {
-        product: {
-          _id,
-        },
-      });
-
-      if (res.data.success) {
-        userDispatch({
-          type: "ADD_TO_CART",
-          payload: {
-            product: _id,
-          },
-        });
-      }
-      setLoading(false);
-      return;
-    }
-
-    navigate("/login", {
-      state: {
-        message: "Login to add to cart.",
-        addTo: "CART",
-        productId: _id,
-      },
-    });
-  };
-
-  const isAlreadyInCart = (_id) =>
-    userData.cartList.some(({ product }) => product === _id);
+  const inStock = product?.stock > 0;
 
   return (
-    <div className="top-margin h-100 flex justify-center align-center">
+    <div className="top-margin h-100 flex place-center">
       <div className="flex border-1 product-detail-wrapper">
         <div className="product-img-wrapper p-6">
           <img
@@ -119,16 +40,14 @@ export const ProductInfo = () => {
             alt={product.name}
           />
         </div>
-        <div className="flex flex-col p-4">
+        <div className="flex-col p-4">
           <div className="heading-2"> {product.name}</div>
           <div className="m-v-8 m-h-4">
             Discounted price:
             <span className="text-bold font-5"> ₹{product.price}</span>
           </div>
-          
-          {!inStock && 
-          <div className="m-4 text-bold">Out of stock!</div>
-          }
+
+          {!inStock && <div className="m-4 text-bold">Out of stock!</div>}
           <div className="flex space-around w-7 m-v-4">
             {isAlreadyInWishList(product._id) ? (
               <button
@@ -163,7 +82,6 @@ export const ProductInfo = () => {
               </button>
             )}
           </div>
-
 
           <div className="heading-3 product-description">Description</div>
           <div className="p-4 font-5">{product.description}</div>
